Issue #2924436 by almaudoh: Convert all SmsFramework WebTests to BrowserTests
