+++ modules/sms_blast/tests/src/Functional/SmsBlastBrowserTest.php
@@ -1,8 +1,8 @@
 <?php
 
-namespace Drupal\sms_blast\Tests;
+namespace Drupal\Tests\sms_blast\Functional;
 
-use Drupal\sms\Tests\SmsFrameworkWebTestBase;
+use Drupal\Tests\sms\Functional\SmsFrameworkBrowserTestBase;
 use Drupal\field\Entity\FieldStorageConfig;
 use Drupal\field\Entity\FieldConfig;
 use Drupal\Component\Utility\Unicode;
@@ -14,7 +14,7 @@ use Drupal\sms\Entity\PhoneNumberSettings;
  * @group SMS Framework
  * @group legacy
  */
-class SmsBlastWebTest extends SmsFrameworkWebTestBase {
+class SmsBlastBrowserTest extends SmsFrameworkBrowserTestBase {
 
   public static $modules = ['sms', 'user', 'sms_blast'];
