+++ tests/src/Functional/SmsFrameworkPhoneNumberTest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Drupal\sms\Tests;
+namespace Drupal\Tests\sms\Functional;
 
 use Drupal\Core\Entity\EntityInterface;
 use Drupal\sms\Message\SmsMessageInterface;
@@ -9,8 +9,9 @@ use Drupal\sms\Message\SmsMessageInterface;
  * General phone number verification tests.
  *
  * @group SMS Framework
+ * @group legacy
  */
-class SmsFrameworkPhoneNumberTest extends SmsFrameworkWebTestBase {
+class SmsFrameworkPhoneNumberTest extends SmsFrameworkBrowserTestBase {
 
   public static $modules = ['entity_test'];
