+++ tests/src/Functional/SmsFrameworkPhoneNumberWidgetTest.php
@@ -1,8 +1,9 @@
 <?php
 
-namespace Drupal\sms\Tests;
+namespace Drupal\Tests\sms\Functional;
 
 use Drupal\Core\Url;
+use Drupal\Tests\Traits\Core\CronRunTrait;
 
 /**
  * Tests phone numbers.
@@ -10,7 +11,9 @@ use Drupal\Core\Url;
  * @group SMS Framework
  * @group legacy
  */
-class SmsFrameworkPhoneNumberWidgetTest extends SmsFrameworkWebTestBase {
+class SmsFrameworkPhoneNumberWidgetTest extends SmsFrameworkBrowserTestBase {
+
+  use CronRunTrait;
 
   public static $modules = ['entity_test'];
