+++ tests/src/Kernel/SmsFrameworkProviderTest.php
@@ -13,7 +13,6 @@ use Drupal\sms\Entity\SmsGateway;
 use Drupal\sms\Direction;
 use Drupal\sms\Event\SmsEvents;
 use Drupal\sms\Message\SmsMessageResultInterface;
-use Drupal\sms\Tests\SmsFrameworkTestTrait;
 
 /**
  * Tests SMS Framework provider service.
